# docs/swagger.yaml
basePath: /
definitions:
  model.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
...
  model.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  model.RegisterRequest:
    properties:
      business_name:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
    required:
    - business_name
    - email
    - name
    - password
    - phone_number
    type: object
  model.TopUpRequest:
    properties:
      amount:
        type: number
    type: object
...
  model.UpdateTopUpRequestBody:
    properties:
      status:
        enum:
        - SUCCESS
        - FAILED
        example: SUCCESS
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/model.ErrorBody'
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      meta: {}
    type: object
host: localhost:8080
info:
  contact: {}
  description: API documentation for Project VioPal
  title: VioPal
  version: "1.0"
paths:
  /api/v1/admin/balance-requests:
    get:
      consumes:
      - application/json
      description: Get paginated list of all merchant top up requests with optional
        filters
      parameters:
      - description: Page number (default 1)
        example: 1
        in: query
        name: page
        type: integer
      - description: Items per page (default 10)
        example: 10
        in: query
        name: perpage
        type: integer
      - description: Top up status
        enum:
        - PENDING
        - SUCCESS
        - FAILED
        in: query
        name: status
        type: string
      - description: Top up reference number
        example: TOPUP-20260504-0001
        in: query
        name: ref_number
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List top up requests (Admin)
      tags:
      - Admin
  /api/v1/admin/balance-requests/{ref_number}/status:
    patch:
      consumes:
      - application/json
      description: Update top up request status by reference number (Admin only)
      parameters:
      - description: Top Up Reference Number
        in: path
        name: ref_number
        required: true
        type: string
      - description: Status only
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateTopUpRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update top up request status
      tags:
      - Admin
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Login using email and password, returns JWT token
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Login user
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Create new merchant user and merchant profile
      parameters:
      - description: Register Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Register merchant account
      tags:
      - Auth
  /api/v1/balance-requests:
    get:
      consumes:
      - application/json
      description: Get paginated list of top up requests for authenticated merchant
        with optional filters
      parameters:
      - description: Page number (default 1)
        example: 1
        in: query
        name: page
        type: integer
      - description: Items per page (default 10)
        example: 10
        in: query
        name: perpage
        type: integer
      - description: Top up status
        enum:
        - PENDING
        - SUCCESS
        - FAILED
        in: query
        name: status
        type: string
      - description: Top up reference number
        example: TOPUP-20260504-0001
        in: query
        name: ref_number
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List top up requests
      tags:
      - Merchant Wallet
      - Merchant Wallet
    post:
      consumes:
      - application/json
      description: Merchant request balance top up
      parameters:
      - description: Top Up Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.TopUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Request top up balance
      tags:
      - Merchant Wallet
...
  /api/v1/me:
    get:
      consumes:
      - application/json
      description: Get profile data by email
      parameters:
      - description: User email
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Auth
...
  /api/v1/wallet:
    get:
      description: Get merchant balance from authenticated user token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get merchant wallet balance
      tags:
      - Merchant Wallet
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"